+++ controller/callbackcontroller.php
@@ -42,6 +42,7 @@ use OCA\Onlyoffice\AppConfig;
 use OCA\Onlyoffice\Crypt;
 use OCA\Onlyoffice\DocumentService;
 use OCA\Onlyoffice\FileVersions;
+use OCA\Onlyoffice\FileUtility;
 use OCA\Onlyoffice\VersionManager;
 use OCA\Onlyoffice\KeyManager;
 use OCA\Onlyoffice\RemoteInstance;
@@ -242,13 +243,24 @@ class CallbackController extends Controller {
 		}
 
 		$shareToken = isset($hashData->shareToken) ? $hashData->shareToken : null;
-		list($file, $error) = empty($shareToken) ? $this->getFile($userId, $fileId, null, $changes ? null : $version, $template) : $this->getFileByToken($fileId, $shareToken, $changes ? null : $version);
+		list($file, $error, $share) = empty($shareToken) ? $this->getFile($userId, $fileId, null, $changes ? null : $version, $template) : $this->getFileByToken($fileId, $shareToken, $changes ? null : $version);
 
 		if (isset($error)) {
 			return $error;
 		}
 
-		if ($this->userSession->isLoggedIn() && !$file->isReadable()) {
+		$canDownload = true;
+
+		$fileStorage = $file->getStorage();
+		if ($fileStorage->instanceOfStorage("\OCA\Files_Sharing\SharedStorage") || !empty($shareToken)) {
+			$share = empty($share) ? $fileStorage->getShare() : $share;
+			$canDownload = FileUtility::canShareDownload($share);
+			if (!$canDownload && !empty($this->config->getDocumentServerSecret())) {
+				$canDownload = true;
+			}
+		}
+
+		if ($this->userSession->isLoggedIn() && !$file->isReadable() || !$canDownload) {
 			$this->logger->error("Download without access right", ["app" => $this->appName]);
 			return new JSONResponse(["message" => $this->trans->t("Access denied")], Http::STATUS_FORBIDDEN);
 		}
@@ -504,7 +516,7 @@ class CallbackController extends Controller {
 						\OC_Util::setupFS($userId);
 					}
 
-					list($file, $error) = empty($shareToken) ? $this->getFile($userId, $fileId, $filePath) : $this->getFileByToken($fileId, $shareToken);
+					list($file, $error, $share) = empty($shareToken) ? $this->getFile($userId, $fileId, $filePath) : $this->getFileByToken($fileId, $shareToken);
 
 					if (isset($error)) {
 						$this->logger->error("track error: $fileId " . json_encode($error->getData()), ["app" => $this->appName]);
@@ -613,7 +625,7 @@ class CallbackController extends Controller {
 	 */
 	private function getFile($userId, $fileId, $filePath = null, $version = 0, $template = false) {
 		if (empty($fileId)) {
-			return [null, new JSONResponse(["message" => $this->trans->t("FileId is empty")], Http::STATUS_BAD_REQUEST)];
+			return [null, new JSONResponse(["message" => $this->trans->t("FileId is empty")], Http::STATUS_BAD_REQUEST), null];
 		}
 
 		try {
@@ -621,12 +633,12 @@ class CallbackController extends Controller {
 			$files = $folder->getById($fileId);
 		} catch (\Exception $e) {
 			$this->logger->logException($e, ["message" => "getFile: $fileId", "app" => $this->appName]);
-			return [null, new JSONResponse(["message" => $this->trans->t("Invalid request")], Http::STATUS_BAD_REQUEST)];
+			return [null, new JSONResponse(["message" => $this->trans->t("Invalid request")], Http::STATUS_BAD_REQUEST), null];
 		}
 
 		if (empty($files)) {
 			$this->logger->error("Files not found: $fileId", ["app" => $this->appName]);
-			return [null, new JSONResponse(["message" => $this->trans->t("Files not found")], Http::STATUS_NOT_FOUND)];
+			return [null, new JSONResponse(["message" => $this->trans->t("Files not found")], Http::STATUS_NOT_FOUND), null];
 		}
 
 		$file = $files[0];
@@ -651,10 +663,10 @@ class CallbackController extends Controller {
 
 			if ($owner !== null) {
 				if ($owner->getUID() !== $userId) {
-					list($file, $error) = $this->getFile($owner->getUID(), $file->getId());
+					list($file, $error, $share) = $this->getFile($owner->getUID(), $file->getId());
 
 					if (isset($error)) {
-						return [null, $error];
+						return [null, $error, null];
 					}
 				}
 
@@ -667,7 +679,7 @@ class CallbackController extends Controller {
 			}
 		}
 
-		return [$file, null];
+		return [$file, null, null];
 	}
 
 	/**
@@ -683,14 +695,14 @@ class CallbackController extends Controller {
 		list($share, $error) = $this->getShare($shareToken);
 
 		if (isset($error)) {
-			return [null, $error];
+			return [null, $error, null];
 		}
 
 		try {
 			$node = $share->getNode();
 		} catch (NotFoundException $e) {
 			$this->logger->logException($e, ["message" => "getFileByToken error", "app" => $this->appName]);
-			return [null, new JSONResponse(["message" => $this->trans->t("File not found")], Http::STATUS_NOT_FOUND)];
+			return [null, new JSONResponse(["message" => $this->trans->t("File not found")], Http::STATUS_NOT_FOUND), null];
 		}
 
 		if ($node instanceof Folder) {
@@ -698,11 +710,11 @@ class CallbackController extends Controller {
 				$files = $node->getById($fileId);
 			} catch (\Exception $e) {
 				$this->logger->logException($e, ["message" => "getFileByToken: $fileId", "app" => $this->appName]);
-				return [null, new JSONResponse(["message" => $this->trans->t("Invalid request")], Http::STATUS_NOT_FOUND)];
+				return [null, new JSONResponse(["message" => $this->trans->t("Invalid request")], Http::STATUS_NOT_FOUND), null];
 			}
 
 			if (empty($files)) {
-				return [null, new JSONResponse(["message" => $this->trans->t("File not found")], Http::STATUS_NOT_FOUND)];
+				return [null, new JSONResponse(["message" => $this->trans->t("File not found")], Http::STATUS_NOT_FOUND), null];
 			}
 			$file = $files[0];
 		} else {
@@ -722,7 +734,7 @@ class CallbackController extends Controller {
 			}
 		}
 
-		return [$file, null];
+		return [$file, null, $share];
 	}
 
 	/**