+++ controller/editorcontroller.php
@@ -141,6 +141,13 @@ class EditorController extends Controller {
 	private $groupManager;
 
 	/**
+	 * Avatar manager
+	 *
+	 * @var IAvatarManager
+	 */
+	private $avatarManager;
+
+	/**
 	 * @param string $AppName - application name
 	 * @param IRequest $request - request object
 	 * @param IRootFolder $root - root folder
@@ -186,6 +193,7 @@ class EditorController extends Controller {
 		$this->versionManager = new VersionManager($AppName, $root);
 
 		$this->fileUtility = new FileUtility($AppName, $trans, $logger, $config, $shareManager, $session);
+		$this->avatarManager = \OC::$server->getAvatarManager();
 	}
 
 	/**
@@ -410,6 +418,49 @@ class EditorController extends Controller {
 					$userElement["email"] = $email;
 				}
 				array_push($result, $userElement);
+			}
+		}
+		return $result;
+	}
+
+	/**
+	 * Get user for Info
+	 *
+	 * @param string $userIds - users identifiers
+	 *
+	 * @return array
+	 *
+	 * @NoAdminRequired
+	 * @NoCSRFRequired
+	 */
+	public function userInfo($userIds) {
+		$result = [];
+		$userIds = json_decode($userIds, true);
+
+		if ($userIds !== null && is_array($userIds)) {
+			foreach ($userIds as $userId) {
+				$userData = [];
+				$user = $this->userManager->get($this->getUserId($userId));
+				if (!empty($user)) {
+					$userData = [
+						"name" => $user->getDisplayName(),
+						"id" => $userId
+					];
+					$avatar = $this->avatarManager->getAvatar($user->getUID());
+					if ($avatar->exists()) {
+						$userAvatarUrl = $this->urlGenerator->getAbsoluteURL(
+							$this->urlGenerator->linkToRoute(
+								"core.avatar.getAvatar",
+								[
+									"userId" => $user->getUID(),
+									"size" => 64,
+								]
+							)
+						);
+						$userData["image"] = $userAvatarUrl;
+					}
+					array_push($result, $userData);
+				}
 			}
 		}
 		return $result;
@@ -761,6 +812,18 @@ class EditorController extends Controller {
 			return ["error" => $this->trans->t("You don't have enough permission to create")];
 		}
 
+		$documentServerUrl = $this->config->getDocumentServerUrl();
+
+		if (empty($documentServerUrl)) {
+			$this->logger->error("documentServerUrl is empty", ["app" => $this->appName]);
+			return ["error" => $this->trans->t("ONLYOFFICE app is not configured. Please contact admin")];
+		}
+
+		if (parse_url($url, PHP_URL_HOST) !== parse_url($documentServerUrl, PHP_URL_HOST)) {
+			$this->logger->error("Incorrect domain in file url", ["app" => $this->appName]);
+			return ["error" => $this->trans->t("The domain in the file url does not match the domain of the Document server")];
+		}
+
 		$url = $this->config->replaceDocumentServerUrlToInternal($url);
 
 		try {
@@ -1092,7 +1155,16 @@ class EditorController extends Controller {
 			$this->logger->error("File for generate presigned url was not found: $dir", ["app" => $this->appName]);
 			return ["error" => $this->trans->t("File not found")];
 		}
-		if (!$file->isReadable()) {
+
+		$canDownload = true;
+
+		$fileStorage = $file->getStorage();
+		if ($fileStorage->instanceOfStorage("\OCA\Files_Sharing\SharedStorage")) {
+			$share = $fileStorage->getShare();
+			$canDownload = FileUtility::canShareDownload($share);
+		}
+
+		if (!$file->isReadable() || !$canDownload) {
 			$this->logger->error("File without permission: $dir", ["app" => $this->appName]);
 			return ["error" => $this->trans->t("You do not have enough permissions to view the file")];
 		}
@@ -1160,6 +1232,14 @@ class EditorController extends Controller {
 		$canDownload = $this->fileUtility->hasPermissionAttribute($file);
 		if (!$canDownload) {
 			return $this->renderError($this->trans->t("Not permitted"));
+		}
+
+		$fileStorage = $file->getStorage();
+		if ($fileStorage->instanceOfStorage("\OCA\Files_Sharing\SharedStorage")) {
+			$share = empty($share) ? $fileStorage->getShare() : $share;
+			if (!FileUtility::canShareDownload($share)) {
+				return $this->renderError($this->trans->t("Not permitted"));
+			}
 		}
 
 		$fileName = $file->getName();
@@ -1458,6 +1538,21 @@ class EditorController extends Controller {
 		}
 
 		return false;
+	}
+
+	/**
+	 * Get Nextcloud userId from unique user identifier
+	 *
+	 * @param string $userId - current user identifier
+	 *
+	 * @return string
+	 */
+	private function getUserId($userId) {
+		if (str_contains($userId, "_")) {
+			$userIdExp = explode("_", $userId);
+			$userId = end($userIdExp);
+		}
+		return $userId;
 	}
 
 	/**