+++ js/editor.js
@@ -273,6 +273,7 @@
         }
 
         OCA.Onlyoffice.resize();
+        OCA.Onlyoffice.setViewport();
     };
 
     OCA.Onlyoffice.onRequestSaveAs = function (event) {
@@ -517,16 +518,31 @@
 
     OCA.Onlyoffice.onRequestUsers = function (event) {
         let operationType = typeof(event.data.c) !== "undefined" ? event.data.c : null;
-        $.get(OC.generateUrl("apps/" + OCA.Onlyoffice.AppName + "/ajax/users?fileId={fileId}&operationType=" + operationType,
-        {
-            fileId: OCA.Onlyoffice.fileId || 0
-        }),
-        function onSuccess(response) {
-            OCA.Onlyoffice.docEditor.setUsers({
-                "c": operationType,
-                "users": response
-            });
-        });
+        switch (operationType) {
+            case "info":
+                    $.get(OC.generateUrl("apps/" + OCA.Onlyoffice.AppName + "/ajax/userInfo?userIds={userIds}",
+                    {
+                        userIds: JSON.stringify(event.data.id)
+                    }),
+                    function onSuccess(response) {
+                        OCA.Onlyoffice.docEditor.setUsers({
+                            "c": operationType,
+                            "users": response
+                        });
+                    });
+                break;
+                default:
+                    $.get(OC.generateUrl("apps/" + OCA.Onlyoffice.AppName + "/ajax/users?fileId={fileId}&operationType=" + operationType,
+                    {
+                        fileId: OCA.Onlyoffice.fileId || 0
+                    }),
+                    function onSuccess(response) {
+                        OCA.Onlyoffice.docEditor.setUsers({
+                            "c": operationType,
+                            "users": response
+                        });
+                    });
+        }
     };
 
     OCA.Onlyoffice.onRequestReferenceData = function (event) {
@@ -690,6 +706,10 @@
                 window.addEventListener("orientationchange", OCA.Onlyoffice.resize);
             }
         }
+    };
+
+    OCA.Onlyoffice.setViewport = function() {
+        document.querySelector('meta[name="viewport"]').setAttribute("content","width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=0");
     };
 
     $(document).ready(OCA.Onlyoffice.InitEditor);