+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - offline viewer for share link
 - updatable list of supported formats
 - filling pdf instead oform
+- fixed guest redirect when limiting the app to groups
 - fixed mobile editor size
 
 ## Added
@@ -15,6 +16,7 @@
 - opening a reference data source
 - changing a reference data source
 - setting for disable editors cron check
+- support of user avatar in editor
 
 ## 8.2.3
 ## Added